workspaces/Waiver_Recon_Agent/backend/requirements.txt
playwright
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
apscheduler
fastapi
uvicorn
thefuzz
pydantic
slack-sdk
pytest
supabase==2.28.3
pytz
